test(refactor): move the testTransformer() helper to separate module
